requirements.txt
kafka-python
boto3
python-dotenv
fastapi
uvicorn